+++ pdb2reaction/workflows/scan_common.py
@@ -413,14 +413,14 @@ def add_scan_common_options(
                     "--zmin",
                     type=float,
                     default=None,
-                    show_default=False,
+                    show_default="the surface minimum",
                     help="Lower bound of color scale for plots (kcal/mol).",
                 ),
                 click.option(
                     "--zmax",
                     type=float,
                     default=None,
-                    show_default=False,
+                    show_default="the surface maximum",
                     help="Upper bound of color scale for plots (kcal/mol).",
                 ),
             ]